requirements.txt
XInput-Python==0.4.0
pyserial==3.4
